Refactor middleware structure: replace morganHelper with morgan middleware, implement JWT handling in jwt.middleware, and enhance logging with Winston; remove obsolete middleware files.
