+++ backend/src/types/cookies.ts
@@ -1,0 +1,8 @@
+import { Request } from "express";
+
+export interface RequestWithCookies extends Request {
+  cookies: {
+    refreshToken?: string;
+    [key: string]: string | undefined;
+  };
+}
